pyproject.toml
[project]
name = "neo4j-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-groq>=0.3.2",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.16",
    "neo4j>=5.28.1",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
]
